Add Spring Boot configuration system with environment variable support
- Add Spring Boot starter parent and dependencies for configuration management
- Convert RuntimeOptions to @ConfigurationProperties bean with prefix 'galeshapley'
- Create application.yml with default configuration values
- Add support for configuration via environment variables (GALESHAPLEY_* prefix)
- Add support for command line arguments (--galeshapley.* format)
- Update Main class to use Spring Boot ApplicationContext
- Add comprehensive documentation in CONFIGURATION.md with usage examples
- Maintain backward compatibility with existing builder pattern
- All tests passing with new Spring Boot configuration system
