+++ src/test/resources/random-large-scale-config.yaml
@@ -46,100 +46,80 @@ simulation:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 42
     m2:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 43
     m3:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 44
     m4:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 45
     m5:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 46
     m6:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 47
     m7:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 48
     m8:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 49
     m9:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 50
     m10:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 51
   proposeePreferences:
     w1:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 52
     w2:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 53
     w3:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 54
     w4:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 55
     w5:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 56
     w6:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 57
     w7:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 58
     w8:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 59
     w9:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 60
     w10:
       generator:
         random: true
         emptySetProbability: 0.2
-        seed: 61
